+++ docs/changes.rst
@@ -6,6 +6,10 @@ Version 0.8.0
 
 To be released.
 
+- Fixed a typing bug that :attr:`Entity.label <wikidata.entity.Entity.label>`
+  and :attr:`Entity.description <wikidata.entity.Entity.description>` properties
+  were incorrectly typed.
+
 
 Version 0.7.0
 -------------